refactor (IN-181): 리뷰 반영 (key 값, import 경로 수정)
