+++ src/entities/channel/kpiCard/ui/KpiCard.tsx
@@ -20,7 +20,7 @@ export function KpiCard({ icon, label, prefix, value, unit }: KpiCardProps) {
         )}
         <span className='text-ibm-title-md-thin'>{label}</span>
       </div>
-      <div className='h-fit w-full pl-[4.4rem] text-text-and-icon-default'>
+      <div className='h-fit w-full pl-[4.4rem] text-text-and-icon-primary'>
         {prefix && <span className='text-ibm-heading-md-normal'>{prefix}</span>}
         <span className='text-ibm-heading-md-normal'>{value}</span>
         <span className='text-ibm-heading-md-normal'>{unit}</span>
